requirements.txt
pydub==0.25.1
openai==1.12.0
python-dotenv==1.0.0
moviepy==1.0.3
requests==2.31.0 